Clock in works without JavaScript, and the workspace has a way out
Research-led portal fixes. Four of these are not polish.

NO SIGN-OUT EXISTED in the employee workspace. Not in the bottom bar, not in the
More drawer, not on Home -- the only one reachable was a link buried on
/portal/profile. A workspace holding payslips, contract terms and loan records, on
shared phones, with no exit. Sign-out is now in the account group and the Home
topbar beside the signed-in name, as a POST: /portal/logout exported GET and POST
with identical bodies, and a GET logout fires on a prefetch, a link scanner or a
third-party image tag.

CLOCK IN AND CLOCK OUT WERE JAVASCRIPT-ONLY. All four punch controls were
type=button with an onclick; the forms had no submit button at all. On a mid-range
Android on a weak connection -- the actual device profile here -- the page renders
and is tappable before the inline script runs, or after it throws. The button does
nothing and the person believes they clocked in. Every other write on that page
posts natively; only the action performed twice a day by everyone was JS-only, and
its failure is invisible to the employee and to HR until a missing punch becomes a
payroll dispute. Now native submit, with the JS as enhancement.

THE BOTTOM BAR HELD THE WRONG THINGS. For an ordinary employee the three slots
resolved to Home, Tasks, Messages -- clock-in, leave and payslips all behind More.
navigation.ts's own comment says checking in is a daily act performed with one
thumb, then assigns it a rank the three-slot slice discards. The comment and the
number disagreed; the number is now fixed.

NINE PORTAL PAGES STILL RENDERED A STATIC FOUR-TAB BAR with no capability test, so
the navigation physically rearranged as a person moved through their own workspace,
and Home pointed somewhere different depending where you stood. WCAG 2.2 SC 3.2.3.
approvals.astro also passed active=hr, so aria-current landed on My HR while the
reader was on Approvals -- a screen-reader user told they are somewhere they are not.

THE FACE MODAL at clock-in had no dialog role, no accessible name, no focus move, no
Escape, and focus stayed on the button behind the scrim. Its 'Skip camera' escape --
the control that lets somebody decline a camera they are entitled to decline, since
the capture is advisory-only -- was 2.6:1 grey at about 20px tall: the least visible
and least tappable element on the screen. That reads as a nudge whether or not it
was designed as one.

astro check: zero in files touched here. Build completes.
NOT VERIFIED: no browser. Every interaction claim above is read from source.

+++ src/lib/announcements.ts
@@ -1394,7 +1394,7 @@ export async function updateAnnouncement(
              ack_required = ${v.ackRequired},
              expires_at = ${v.expiresAt}::timestamptz,
              links = ${v.linksJson}::jsonb,
-             version = version + ${wasPublished ? 1 : 0},
+             version = version + (${wasPublished ? 1 : 0})::int,
              updated_at = NOW()
        WHERE id = ${id}::uuid`);
 
@@ -1794,7 +1794,7 @@ export async function upcomingMilestones(viewer: FeedViewer, days = 7): Promise<
            to_char(e.date_of_birth, 'MM-DD') = ANY (
              SELECT to_char(gs, 'MM-DD')
                FROM generate_series(NOW() - INTERVAL '2 days',
-                                    NOW() + (${window} * INTERVAL '1 day'),
+                                    NOW() + ((${window})::int * INTERVAL '1 day'),
                                     INTERVAL '1 day') AS gs
            )
          )
@@ -1827,7 +1827,7 @@ export async function upcomingMilestones(viewer: FeedViewer, days = 7): Promise<
            to_char(e.joining_date, 'MM-DD') = ANY (
              SELECT to_char(gs, 'MM-DD')
                FROM generate_series(NOW() - INTERVAL '2 days',
-                                    NOW() + (${window} * INTERVAL '1 day'),
+                                    NOW() + ((${window})::int * INTERVAL '1 day'),
                                     INTERVAL '1 day') AS gs
            )
          )
